src_py3/requirements.txt
Flask
SpeechRecognition
google-cloud-speech
google-auth
whisper-ai
ollama
torch
torchvision
torchaudio